requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-openai
pymupdf
pypdf
flask
pickle
re

faiss-cpu
numpy